re-structuring project with atomic design pattern and separating some components
