+++ src/components/atoms/CardBody.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { cn } from '../../utils/utils';
+
+export const CardBody = ({ children, className }: { children: React.ReactNode; className?: string }) => {
+  return (
+    <div className={cn('h-96 w-96 [transform-style:preserve-3d] [&>*]:[transform-style:preserve-3d]', className)}>
+      {children}
+    </div>
+  );
+};
