+++ src/components/atoms/MouseEnterContext.tsx
@@ -1,0 +1,21 @@
+import React, { createContext, useContext, useState } from 'react';
+
+const MouseEnterContext = createContext<[boolean, React.Dispatch<React.SetStateAction<boolean>>] | undefined>(undefined);
+
+export const MouseEnterProvider = ({ children }: { children: React.ReactNode }) => {
+  const [isMouseEntered, setIsMouseEntered] = useState(false);
+
+  return (
+    <MouseEnterContext.Provider value={[isMouseEntered, setIsMouseEntered]}>
+      {children}
+    </MouseEnterContext.Provider>
+  );
+};
+
+export const useMouseEnter = () => {
+  const context = useContext(MouseEnterContext);
+  if (context === undefined) {
+    throw new Error('useMouseEnter must be used within a MouseEnterProvider');
+  }
+  return context;
+};
