+++ src/components/organisms/Box.tsx
@@ -1,4 +1,4 @@
-import Card from './card';
+import Card from "../molecules/Card";
 
 interface BoxProps {
   startColumn?: number;
