+++ src/components/organisms/SeeResume.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { useIsMobile } from '../hooks/useIsMobile';
-import ResumePDF from '../assets/George Gil Garib Resume.pdf';
-import CVPDF from '../assets/George Gil Garib CV.pdf';
-import resumeIcon from '../assets/resume-icon.svg';
-import { useLanguage } from '../hooks/useLanguage';
+import { useIsMobile } from '../../hooks/useIsMobile';
+import ResumePDF from '../../assets/documents/George Gil Garib Resume.pdf'
+import CVPDF from '../../assets/documents/George Gil Garib CV.pdf';
+import resumeIcon from '../../assets/icons/resume-icon.svg';
+import { useLanguage } from '../../hooks/useLanguage';
 import { useTranslation } from 'react-i18next';
 
 export default function SeeResume(): React.ReactElement {
