--- src/layout/__test__/Nabbar.spec.ts
@@ -1,16 +1,0 @@
-// import React from'react';
-// import { render } from'@testing-library/react';
-// import { BrowserRouter as Router } from'react-router-dom';
-// import Navbar from'../Navbar.tsx';
-// import { describe, it, expect } from'vitest';
-
-// describe('Navbar Component', () => {
-//   it('should render Navbar correctly and match snapshot', () => {
-//     // const { asFragment } = render(
-//     //   <Router><Navbar /></Router>
-//     // );
-//     // expect(asFragment()).toMatchSnapshot();
-//     expect(1).toBe(1)
-//     console.log('si');
-//   });
-// });
