+++ src/locales/en.local.ts
@@ -11,8 +11,7 @@ export const en = {
   music: {
     favoritePlaylist: {
       title: 'My Favorite Playlist',
-      description:
-        'Enjoy my personal selection of songs in this playlist. Listen to it and let yourself be carried away by the rhythm!',
+      description: 'Enjoy my personal selection of songs in this playlist. Listen to it and let yourself be carried away by the rhythm!',
     },
     video: {
       clickMe: 'Click me!',
@@ -21,19 +20,19 @@ export const en = {
   },
   me: {
     presentation: {
-        title: "Hi! I'm George Gil",
-        description: "I'm a curious software engineer who also plays guitar, not always at the same time though.",
+      title: "Hi! I'm George Gil",
+      description: "I'm a curious software engineer who also plays guitar, not always at the same time though.",
     },
     jokes: {
-        joke1: "I decided to create this presentation because, let's face it, my jokes aren’t as good as my code... but here we are.",
-        joke2: "Honestly, putting together a page about myself is way less awkward than trying to have a conversation with myself.",
-        joke3: "In theory, I’m an expert... but since I’m better at doing than explaining, I figured it’s best to just show you.",
-        joke4: "Some people make presentations to impress, I made mine to avoid talking... (didn't work anyway)."        
-    }
+      joke1: "I decided to create this presentation because, let's face it, my jokes aren’t as good as my code... but here we are.",
+      joke2: 'Honestly, putting together a page about myself is way less awkward than trying to have a conversation with myself.',
+      joke3: 'In theory, I’m an expert... but since I’m better at doing than explaining, I figured it’s best to just show you.',
+      joke4: "Some people make presentations to impress, I made mine to avoid talking... (didn't work anyway).",
     },
-    dev: {
-        resume: {
-            seeResume: 'See resume'
-        }
-    }
+  },
+  dev: {
+    resume: {
+      seeResume: 'See resume',
+    },
+  },
 };
