+++ src/locales/es.local.ts
@@ -11,8 +11,7 @@ export const es = {
   music: {
     favoritePlaylist: {
       title: 'Mi Playlist Favorita',
-      description:
-        'Disfruta de mi selección personal de canciones en esta playlist. ¡Escúchala y déjate llevar por el ritmo!',
+      description: 'Disfruta de mi selección personal de canciones en esta playlist. ¡Escúchala y déjate llevar por el ritmo!',
     },
     video: {
       clickMe: 'Clickeame!',
@@ -22,19 +21,18 @@ export const es = {
   me: {
     presentation: {
       title: '¡Hola! Soy George Gil',
-      description:
-        'Soy un curioso ingeniero de softwre que también toca la guitarra, aunque no siempre al mismo tiempo.',
+      description: 'Soy un curioso ingeniero de softwre que también toca la guitarra, aunque no siempre al mismo tiempo.',
     },
     jokes: {
       joke1: 'Decidí hacer esta presentación porque, siendo sinceros, mis chistes no son tan buenos como mi código... pero aquí estamos.',
-      joke2: "A decir verdad, crear una página sobre mí mismo es mucho menos incómodo que tener una conversación conmigo mismo.",
-      joke3: "En teoría, soy un experto... en la práctica, como soy mejor haciendo que explicando, mejor les muestro.",
-      joke4: "Algunos hacen presentaciones para impresionar, yo la hice para evitar hablar... aunque tampoco lo conseguí del todo."
+      joke2: 'A decir verdad, crear una página sobre mí mismo es mucho menos incómodo que tener una conversación conmigo mismo.',
+      joke3: 'En teoría, soy un experto... en la práctica, como soy mejor haciendo que explicando, mejor les muestro.',
+      joke4: 'Algunos hacen presentaciones para impresionar, yo la hice para evitar hablar... aunque tampoco lo conseguí del todo.',
     },
   },
   dev: {
     resume: {
-      seeResume: 'Ver CV'
-    }
-  }
+      seeResume: 'Ver CV',
+    },
+  },
 };
