+++ src/store/hash/types.ts
